+++ src/interfaces.ts
@@ -8,4 +8,5 @@ export interface Options {
   import?: { module: string; export: string };
   tag?: string;
   attributes?: Array<OptionAttrMap | { preset?: 'lit-html' | 'global' }>;
+  define?: string;
 }