+++ CHANGELOG.md
@@ -19,9 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.0] - 2020-08-01
 
-### Unreleased
-Convert JSX String Literal properties to the values of attributes on the specified elements.
+### Added
 Add an option to set a custom function name for define Custom Element.
+### Changed
+Convert JSX String Literal properties to the values of attributes on the specified elements.
 
 
 ## [0.0.0] - 2020-04-16